refactor: landing 페이지를 index 로 변경
